chore: consolidate duplicated API infrastructure into api.types.ts and applications.api.ts
