+++ ui/catalog/src/services/auth.ts
@@ -2,7 +2,7 @@ import { api } from "@/api/axios";
 import { AUTH_ENDPOINTS } from "@/constants/api-endpoints.constants";
 import { useAuthStore } from "@/store/auth.store";
 import { useDeployStore } from "@/store/deploy.store";
-import { fetchArchitectures } from "@/api/digitalAssistants";
+import { fetchArchitectures } from "@/api/applications.api";
 import type { LoginRequest, LoginResponse, UserInfo } from "@/types/auth";
 import { useServiceDeployStore } from "@/store/serviceDeploy.store";
