+++ ui/catalog/src/utils/serviceDeploymentTransform.ts
@@ -2,11 +2,8 @@ import type {
   DeployFormData,
   ComponentConfig,
 } from "@/components/ServicesDeployFlow/types";
-import type {
-  ServiceDeployOptions,
-  ProviderSchema,
-} from "@/services/deployment.api";
-import { fetchProviderParams } from "@/services/deployment.api";
+import type { ServiceDeployOptions, ProviderSchema } from "@/types/api.types";
+import { fetchProviderParams } from "@/api/applications.api";
 
 interface DeploymentComponent {
   component_type: string;
